requirements.txt
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
boto3==1.26.137
botocore==1.29.137
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.1
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
inflect==7.5.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.8.0
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.31.0
resend==0.8.0
rsa==4.9.1
s3transfer==0.6.2
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.1.1
SQLAlchemy==2.0.41
starlette==0.49.3
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.18
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
